Refactor transport and server grpcs into BundleServer and BundleExchange (#273)
We want to unify the sending and receiving of bundles so that the same
interface is used on both the BundleServer and the BundleTransport.
BundleServer now has a single method to tell the transport what bundles
should be exchanged.

This moved the core DDD code from BundleClient to bundle-core so that we can do end-to-end tests from client to server.

+++ bundle-core/src/main/java/net/discdd/bundlesecurity/ServerSecurity.java
@@ -213,15 +213,18 @@ public class ServerSecurity {
         clientDataPath.toFile().mkdirs();
         logger.log(FINE, "[ServerSecurity]:Client Data Path = " + clientDataPath);
         try {
-            Files.copy(clientKeyPath.resolve(SecurityUtils.CLIENT_IDENTITY_KEY),
-                       clientDataPath.resolve(SecurityUtils.CLIENT_IDENTITY_KEY));
-            Files.copy(clientKeyPath.resolve(SecurityUtils.CLIENT_BASE_KEY),
-                       clientDataPath.resolve(SecurityUtils.CLIENT_BASE_KEY));
+            if (Files.notExists(clientDataPath.resolve(SecurityUtils.CLIENT_IDENTITY_KEY))) {
+                Files.copy(clientKeyPath.resolve(SecurityUtils.CLIENT_IDENTITY_KEY),
+                           clientDataPath.resolve(SecurityUtils.CLIENT_IDENTITY_KEY));
+            }
+            if (Files.notExists(clientDataPath.resolve(SecurityUtils.CLIENT_BASE_KEY))) {
+                Files.copy(clientKeyPath.resolve(SecurityUtils.CLIENT_BASE_KEY),
+                           clientDataPath.resolve(SecurityUtils.CLIENT_BASE_KEY));
+            }
         } catch (IOException e) {
             logger.log(SEVERE,
                        "[ServerSecurity] INFO: Client Keys already exist Client Data Path for client ID " + clientID,
                        e);
-            logger.log(SEVERE, "[SEC] INFO: Client Keys already exist Client Data Path for client ID " + clientID);
         }
 
         initializeClientKeysFromFiles(clientDataPath, clientSession);
